move onboarding widgets to widgets folder
